+++ src/main/java/pe/com/msif/dto/RecommendationDto.java
@@ -4,10 +4,9 @@ import lombok.Data;
 
 @Data
 public class RecommendationDto {
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
-    private Integer medicalHistoryId;
+    private Long medicalHistoryId;
     private Boolean isActive;
 }
-
