+++ src/main/java/pe/com/msif/repository/MedicalHistoryRepository.java
@@ -7,7 +7,6 @@ import pe.com.msif.model.MedicalHistory;
 import java.util.List;
 
 @Repository
-public interface MedicalHistoryRepository extends JpaRepository<MedicalHistory, Integer> {
+public interface MedicalHistoryRepository extends JpaRepository<MedicalHistory, Long> {
     List<MedicalHistory> findAllByIsActive(Boolean isActive);
 }
-
